Tree

+++ Scripts/MeshArea.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 public partial class MeshArea : Area3D
 {
     [Signal]
-    public delegate void AreaSelectedEventHandler(float row, float col);
+    public delegate void AreaSelectedEventHandler(int row, int col);
 
     public void OnAreaInputEvent(Node camera, InputEvent e, Vector3 pos, Vector3 normal, int shape){
        if(e is InputEventMouseButton click && click.IsPressed()){
